backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
defusedxml==0.7.1
django-extra-fields==3.0.2
django-templated-mail==1.1.1
Django==3.2
djangorestframework-simplejwt==5.3.0
djangorestframework==3.12.4
djoser==2.2.0
idna==3.4
oauthlib==3.2.2
Pillow==10.0.1
pip==23.2.1
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
requests==2.31.0
setuptools==58.0.4
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing-extensions==4.8.0
urllib3==1.26.6
